requirements.txt
asyncblink
appdirs
inflect
irc
jaraco.classes
jaraco.collections
jaraco.functools
jaraco.itertools
jaraco.logging
jaraco.stream
jaraco.text
matrix-client
more-itertools
packaging
pyparsing
pytz
PyYAML
requests
six
tempora
git+https://github.com/Cadair/python-appservice-framework.git
git+https://github.com/Cadair/matrix-python-sdk.git@user_id2
future-fstrings
sqlalchemy
pyCrypto